refactor: appContainer 적용
