+++ app/src/main/java/woowacourse/shopping/ui/screens/product/ProductViewModel.kt
@@ -130,10 +130,10 @@ class ProductViewModel(
             initializer {
                 val app = this[APPLICATION_KEY] as ShoppingApplication
                 ProductViewModel(
-                    productRepository = app.productRepository,
-                    cartRepository = app.cartRepository,
-                    recentProductRepository = app.recentProductRepository,
-                    networkMonitor = app.networkMonitor,
+                    productRepository = app.appContainer.productRepository,
+                    cartRepository = app.appContainer.cartRepository,
+                    recentProductRepository = app.appContainer.recentProductRepository,
+                    networkMonitor = app.appContainer.networkMonitor,
                 )
             }
         }
